+++ AetherBags/changelog.md
@@ -1,6 +1,7 @@
 # 1.0.0.3
 - Add options to IPC to access the AetherBags configuration and make changes.
 - Add configurable keybinds for focusing search and opening bag windows directly into search.
+- Added custom per category sorting options.
 
 # 1.0.0.2
 - Fix ghost hovers for expanded retainer grid too.